refactored TableHeader component
